+++ src/Record.js
@@ -120,6 +120,8 @@ const previousPath = 'previous'
  * @param {Object} [opts] Configuration options.
  * @param {boolean} [opts.noValidate=false] Whether to skip validation on the
  * initial properties.
+ * @param {boolean} [opts.validateOnSet=true] Whether to enable setter
+ * validation on properties after the Record has been initialized.
  * @since 3.0.0
  */
 function Record (props, opts) {
@@ -129,9 +131,7 @@ function Record (props, opts) {
   opts || (opts = {})
   const _set = this._set
   _set(creatingPath, true)
-  if (opts.noValidate) {
-    _set(noValidatePath, opts.noValidate === undefined ? true : opts.noValidate)
-  }
+  _set(noValidatePath, !!opts.noValidate)
   _set(keepChangeHistoryPath, opts.keepChangeHistory === undefined ? (mapper ? mapper.keepChangeHistory : true) : opts.keepChangeHistory)
 
   // Set the idAttribute value first, if it exists.
@@ -143,8 +143,13 @@ function Record (props, opts) {
 
   utils.fillIn(this, props)
   _set(creatingPath, false)
-  const validateOnSet = opts.validateOnSet === undefined ? (mapper ? mapper.validateOnSet : true) : opts.validateOnSet
-  _set(noValidatePath, !validateOnSet)
+  if (opts.validateOnSet !== undefined) {
+    _set(noValidatePath, !opts.validateOnSet)
+  } else if (mapper && mapper.validateOnSet !== undefined) {
+    _set(noValidatePath, !mapper.validateOnSet)
+  } else {
+    _set(noValidatePath, false)
+  }
   _set(previousPath, mapper ? mapper.toJSON(props) : utils.plainCopy(props))
 }
 
@@ -844,6 +849,11 @@ export default Component.extend({
   validate (opts) {
     return this._mapper().validate(this, opts)
   }
+}, {
+  creatingPath,
+  noValidatePath,
+  keepChangeHistoryPath,
+  previousPath
 })
 
 /**